+++ src/pages/workspaces/workspace/tools/WorkflowView.js
@@ -3,7 +3,7 @@ import { Fragment } from 'react'
 import { div, h, p, span } from 'react-hyperscript-helpers'
 import { AutoSizer } from 'react-virtualized'
 import * as breadcrumbs from 'src/components/breadcrumbs'
-import { buttonPrimary, buttonSecondary, link, spinnerOverlay, tooltip } from 'src/components/common'
+import { buttonPrimary, buttonSecondary, link, Select, spinnerOverlay, tooltip } from 'src/components/common'
 import { centeredSpinner, icon } from 'src/components/icons'
 import { AutocompleteTextInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
@@ -17,7 +17,7 @@ import * as Nav from 'src/libs/nav'
 import * as StateHistory from 'src/libs/state-history'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
-import { Component, Select } from 'src/libs/wrapped-components'
+import { Component } from 'src/libs/wrapped-components'
 import LaunchAnalysisModal from 'src/pages/workspaces/workspace/tools/LaunchAnalysisModal'
 import WorkspaceContainer from 'src/pages/workspaces/workspace/WorkspaceContainer'
